+++ classes/CatList.class.php
@@ -5,7 +5,7 @@
 *   @author     Lee Garner <lee@leegarner.com>
 *   @copyright  Copyright (c) 2016-2017 Lee Garner <lee@leegarner.com>
 *   @package    classifieds
-*   @version    1.1.3
+*   @version    1.2.1
 *   @license    http://opensource.org/licenses/gpl-2.0.php 
 *               GNU Public License v2 or later
 *   @filesource
@@ -54,14 +54,15 @@ class CatList
         $T->set_file('page', 'catlist.thtml');
 
         // Get all the root categories
-        $sql = "SELECT * FROM {$_TABLES['ad_category']} 
+        /*$sql = "SELECT * FROM {$_TABLES['ad_category']} 
                 WHERE papa_id = 0 " . COM_getPermSQL('AND', 0, 2) .
                 " ORDER BY cat_name ASC";
         $cats = DB_query($sql);
-        if (!$cats) return CLASSIFIEDS_errorMsg($LANG_ADVT['database_error'], 'alert');
-
+        if (!$cats) return CLASSIFIEDS_errorMsg($LANG_ADVT['database_error'], 'alert');*/
+        $Cats = Category::SubCats();
         // If no root categories exist, display just return a message
-        if (DB_numRows($cats) == 0) {
+        //if (DB_numRows($cats) == 0) {
+        if (count($Cats) == 0) {
             $T->set_var('no_cat_found', 
                 "<p align=\"center\" class=\"headmsg\">
                 $LANG_ADVT[no_cat_found]</p>\n");
@@ -72,18 +73,21 @@ class CatList
         $T->set_block('page', 'CatRows', 'CRow');
 
         $i = 1;
-        while ($catsrow = DB_fetchArray($cats)) {
+        $newtime = time() - 3600 * 24 * $_CONF_ADVT['newcatdays'];
+        //while ($catsrow = DB_fetchArray($cats)) {
+        foreach ($Cats as $Cat) {
             // For each category, find the total ad count (including subcats)
             // and display the subcats below it.
             $T->set_var(array(
                 'rowstart'  => $i % 2 == 1 ? "<tr>\n" : '',
-                'cat_url'   => CLASSIFIEDS_makeUrl('home', $catsrow['cat_id']),
-                'cat_name'  => $catsrow['cat_name'],
-                'cat_ad_count' => Category::TotalAds($catsrow['cat_id']),
-                'image' => $catsrow['image'] ? Category::thumbUrl($catsrow['image']) : '',
+                'cat_url'   => CLASSIFIEDS_makeUrl('home', $Cat->cat_id),
+                'cat_name'  => $Cat->cat_name,
+                'cat_ad_count' => Category::TotalAds($Cat->cat_id),
+                'image' => $Cat->image ? Category::thumbUrl($Cat->image) : '',
             ) );
 
-            $sql = "SELECT * FROM {$_TABLES['ad_category']} 
+            $SubCats = Category::SubCats($Cat->cat_id);
+            /*$sql = "SELECT * FROM {$_TABLES['ad_category']} 
                     WHERE papa_id={$catsrow['cat_id']} " .
                         COM_getPermSQL('AND', 0, 2) . "
                     ORDER BY cat_name ASC";
@@ -91,35 +95,31 @@ class CatList
             $subcats = DB_query($sql);
             if (!$subcats) return CLASSIFIEDS_errorMsg($LANG_ADVT['database_error'], 'alert');
 
-            $num = DB_numRows($subcats);
-            $time = time();
+            $num = DB_numRows($subcats);*/
+            $num = count($SubCats);
             // Earliest time to be considered "new"
-            $newtime = $time - 3600 * 24 * $_CONF_ADVT['newcatdays'];
             $subcatlist = '';
 
             $j = 1;
-            while ($subcatsrow = DB_fetchArray($subcats)) {
-                $isnew = $subcatsrow['add_date'] > $newtime ? 
+            //while ($subcatsrow = DB_fetchArray($subcats)) {
+            foreach ($SubCats as $SubCat) {
+                $isnew = $SubCat->add_date > $newtime ? 
                     "<img src=\"{$_CONF['site_url']}/{$_CONF_ADVT['pi_name']}/images/new.gif\" align=\"top\">" : '';
-
                 $subcatlist .= '<a href="'.
-                        CLASSIFIEDS_makeURL('home', $subcatsrow['cat_id']). '">'.
-                        "{$subcatsrow['cat_name']}</a>&nbsp;(" .
-                        Category::TotalAds($subcatsrow['cat_id']). ")&nbsp;{$isnew}";
+                        CLASSIFIEDS_makeURL('home', $SubCat->cat_id). '">'.
+                        "{$SubCat->cat_name}</a>&nbsp;(" .
+                        Category::TotalAds($SubCat->cat_id). ")&nbsp;{$isnew}";
 
                 if ($num != $j)
                     $subcatlist .= ", ";
 
                 $j++;
             }
-
             $T->set_var('subcatlist', $subcatlist);
             $T->set_var('rowend', $i % 2 == 0 ? "</tr>\n" : "");
             $i++;
-
             $T->parse('CRow', 'CatRows', true);
         }
-
         $T->parse('output', 'page');
         return $T->finish($T->get_var('output'));
     }
@@ -139,16 +139,17 @@ class CatList
         $T->set_file('page', 'catlist_blocks.thtml');
 
         // Get all the root categories
-        $sql = "SELECT * FROM {$_TABLES['ad_category']} 
-                WHERE papa_id='' " .
+        /*$sql = "SELECT * FROM {$_TABLES['ad_category']} 
+                WHERE papa_id = 1 " .
                     COM_getPermSQL('AND', 0, 2) .
                 " ORDER BY cat_name ASC";
         //echo $sql;die;
         $cats = DB_query($sql);
         if (!$cats) return CLASSIFIEDS_errorMsg($LANG_ADVT['database_error'], 'alert');
-
+        */
+        $Cats = Category::SubCats();
         // If no root categories exist, display just return a message
-        if (DB_numRows($cats) == 0) {
+        if (count($Cats) == 0) {
             $T->set_var('no_cat_found', 
                 "<p align=\"center\" class=\"headmsg\">
                 $LANG_ADVT[no_cat_found]</p>\n");
@@ -159,16 +160,17 @@ class CatList
         $max = count($CatListcolors);
 
         $i = 0;
-        while ($catsrow = DB_fetchArray($cats)) {
-            if ($catsrow['fgcolor'] == '' || $catsrow['bgcolor'] == '') {
+        //while ($catsrow = DB_fetchArray($cats)) {
+        foreach ($Cats as $Cat) {
+            if ($Cat->fgcolor == '' || $Cat->bgcolor == '') {
                 if ($i >= $max) $i = 0;
                 $bgcolor = $CatListcolors[$i][0];
                 $fgcolor = $CatListcolors[$i][1];
                 $hdcolor = $CatListcolors[$i][2];
                 $i++;
             } else {
-                $fgcolor = $catsrow['fgcolor'];
-                $bgcolor = $catsrow['bgcolor'];
+                $fgcolor = $Cat->fgcolor;
+                $bgcolor = $Cat->bgcolor;
             } 
 
             // For each category, find the total ad count (including subcats)
@@ -177,18 +179,18 @@ class CatList
             $T->set_var(array(
                 'bgcolor'   => $bgcolor,
                 'fgcolor'   => $fgcolor,
-                'cat_url'   => CLASSIFIEDS_makeUrl('home',$catsrow['cat_id']),
-                'cat_name'  => $catsrow['cat_name'],
-                'cat_desc'  => $catsrow['description'],
-                'cat_ad_count' => Category::TotalAds($catsrow['cat_id']),
-                'image' => $catsrow['image'] ? Category::thumbUrl($catsrow['image']): '',
+                'cat_url'   => CLASSIFIEDS_makeUrl('home', $Cat->cat_id),
+                'cat_name'  => $Cat->cat_name,
+                'cat_desc'  => $Cat->description,
+                'cat_ad_count' => Category::TotalAds($Cat->cat_id),
+                'image' => Category::thumbUrl($Cat->image),
             ) );
             $T->parse('Div', 'CatDiv', true);
         }
-
         $T->parse('output', 'page');
         return $T->finish($T->get_var('output'));
     }
+
 }
 
 ?>